+++ mqm-client/mqm-rest-client/src/main/java/com/hp/mqm/client/model/TaxonomyType.java
@@ -4,15 +4,15 @@ package com.hp.mqm.client.model;
 
 final public class TaxonomyType {
 
-    final private int id;
+    final private Long id;
     final private String name;
 
-    public TaxonomyType(int id, String name) {
+    public TaxonomyType(Long id, String name) {
         this.id = id;
         this.name = name;
     }
 
-    public int getId() {
+    public Long getId() {
         return id;
     }
